注册、登陆

+++ src/main/java/com/jifen/manage/demo/service/FunctionService.java
@@ -19,5 +19,5 @@ public interface FunctionService {
      * @param id
      * @return
      */
-    List<FunctionEntity> getFunctionsById(int id);
+    List<FunctionEntity> getFunctionsById(Long id);
 }
